Change ClipAmmo from float type to int32 type using FGCInt32PropertyWrapper

+++ ShooterTest/Source/ShooterTest/Private/Inventory/Item/Gun/STGameplayAbility_FireGun.cpp
@@ -232,7 +232,7 @@ bool USTGameplayAbility_FireGun::CheckCost(const FGameplayAbilitySpecHandle Hand
 	}
 
 	// If we don't have enough ammo
-	const float ClipAmmoAfterNextShot = ClipAmmoSubobject->ClipAmmo - AmmoCost;
+	const int32 ClipAmmoAfterNextShot = ClipAmmoSubobject->ClipAmmo - AmmoCost;
 	if (ClipAmmoAfterNextShot < 0)
 	{
 		UE_LOG(LogSTGameplayAbility, Verbose, TEXT("%s() Not enough ammo to perform a fire. returned false"), ANSI_TO_TCHAR(__FUNCTION__));
